Fix compilation issues.
I ran into a bunch of issues trying to compile this with Ninja on
Linux (yes I know you have some incompatible dependencies, but
ignoring those for now).

The members named the same as struct defs, and __forceinline were the
main issues. Aside from that, it was the default initializer magick on
Mix (I don't kno if there's a standard way to do what you want here).

__forceinline is all over the place in the submodules, so probably
it'd be better to override it with a macro def in the build system.

There were also a couple of incompatible paths in #includes that
needed fixing.