New engine from BMW M3 E46
I was just messing around with the M52B28 and 2JZ a little and settled on this monstrosity of an engine. It's supposed to sound like a BMW M3 E46 and originate from the same vehicle. I then managed to get it to run and sound like the real thing.